# api/mainservice/swagger.yaml
basePath: /
definitions:
...
  internal_handlers_user.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  keycloak.ErrorResponse:
    properties:
      error:
        type: string
      error_description:
        type: string
    type: object
  keycloak.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  keycloak.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  keycloak.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  keycloak.RegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      firstName:
        maxLength: 255
        type: string
      lastName:
        maxLength: 255
        type: string
      middleName:
        maxLength: 255
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    type: object
  keycloak.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_expires_in:
        type: integer
      refresh_token:
        type: string
      scope:
        type: string
      token_type:
        type: string
    type: object
  user.CreateRequest:
    properties:
      email:
        maxLength: 255
        type: string
      firstName:
        maxLength: 255
        type: string
      lastName:
        maxLength: 255
        type: string
      middleName:
        maxLength: 255
        type: string
      uuid:
        type: string
    required:
    - email
    - uuid
    type: object
  user.Response:
    properties:
      createdAt:
        type: string
      deleted:
        type: boolean
      deletedAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      middleName:
        type: string
      role:
        type: string
      uuid:
        type: string
    type: object
host: fishouk-otus-ms.ru
info:
  contact:
    name: OtusMS Support
    url: https://github.com/FischukSergey/otus-ms
  description: |-
    Введите: Bearer {токен} — с префиксом и пробелом.
    Токен: POST /api/v1/auth/login. Роли в claim realm_access.roles
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OtusMS Main Service API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: |-
        Аутентифицирует пользователя и возвращает access и refresh токены.
        Токены получаются от Keycloak.
      parameters:
      - description: Учетные данные пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/keycloak.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Токены успешно получены
          schema:
            $ref: '#/definitions/keycloak.TokenResponse'
        "400":
          description: Невалидный запрос или отсутствуют обязательные поля
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера или ошибка Keycloak
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
      summary: Вход в систему
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: |-
        Инвалидирует refresh токен в Keycloak, завершая сессию пользователя.
        После logout токены становятся недействительными.
      parameters:
      - description: Refresh токен для инвалидации
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/keycloak.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Выход выполнен успешно
        "400":
          description: Невалидный запрос или отсутствует refresh token
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
        "500":
          description: Ошибка при выполнении logout в Keycloak
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
      summary: Выход из системы
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: |-
        Обменивает refresh токен на новую пару access и refresh токенов.
        Позволяет продлить сессию пользователя без повторного ввода пароля.
      parameters:
      - description: Refresh токен
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/keycloak.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Новые токены успешно получены
          schema:
            $ref: '#/definitions/keycloak.TokenResponse'
        "400":
          description: Невалидный запрос или отсутствует refresh token
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
        "401":
          description: Невалидный или истекший refresh токен
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера или ошибка Keycloak
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
      summary: Обновление токена
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: |-
        Создаёт пользователя в Keycloak и Main Service с ролью user.
        После регистрации необходимо выполнить login для получения токенов.
      parameters:
      - description: Данные для регистрации
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/keycloak.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
        "400":
          description: Невалидные данные или ошибка валидации
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
        "409":
          description: Пользователь с таким email уже существует
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/keycloak.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - auth
  /api/v1/users:
    get:
      description: Возвращает всех пользователей из таблицы users, включая мягко удалённых.
        Требуется роль admin.
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей
          schema:
            items:
              $ref: '#/definitions/user.Response'
            type: array
        "401":
          description: Не авторизован - отсутствует или невалидный JWT токен
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "403":
          description: Доступ запрещён - требуется роль admin
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список всех пользователей (только admin)
      tags:
      - users
    post:
      consumes:
      - application/json
      description: |-
        Создаёт нового пользователя. UUID и email должны быть уникальными.
        Доступно для роли admin или service-account (для service-to-service
        коммуникации между Auth-Proxy и Main Service).
      parameters:
      - description: Данные нового пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Невалидный запрос или ошибка валидации
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "401":
          description: Не авторизован - отсутствует или невалидный JWT токен
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "403":
          description: Доступ запрещён - требуется роль admin или service-account
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать пользователя (admin или service account)
      tags:
      - users
  /api/v1/users/{uuid}:
    delete:
      description: Мягкое удаление (soft delete). Запись остаётся в БД с флагом deleted=true.
        Требуется роль admin.
      parameters:
      - description: 'UUID пользователя (формат: uuid4)'
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Пользователь успешно удалён
        "400":
          description: Невалидный UUID
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "401":
          description: Не авторизован - отсутствует или невалидный JWT токен
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "403":
          description: Доступ запрещён - недостаточно прав (требуется роль admin)
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить пользователя (только admin)
      tags:
      - users
    get:
      description: Возвращает данные пользователя по UUID. Мягко удалённый вернётся
        с deleted=true. Требуется роль admin.
      parameters:
      - description: 'UUID пользователя (формат: uuid4)'
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/user.Response'
        "400":
          description: Невалидный UUID
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "401":
          description: Не авторизован - отсутствует или невалидный JWT токен
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "403":
          description: Доступ запрещён - недостаточно прав (требуется роль admin)
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_handlers_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить пользователя (только admin)
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"